+++ qubes/api/internal.py
@@ -29,6 +29,8 @@ import qubes.api.admin
 import qubes.vm.adminvm
 import qubes.vm.dispvm
 
+QUBESD_INTERNAL_SOCK = '/var/run/qubesd.internal.sock'
+
 
 class QubesInternalAPI(qubes.api.AbstractQubesAPI):
     ''' Communication interface for dom0 components,
